requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.1.0
azure-ai-documentintelligence==1.0.0
azure-core==1.32.0
backoff==2.2.1
beautifulsoup4==4.12.3
build==0.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.0.12
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.17
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
Django==5.1.5
djangorestframework==3.15.2
document==1.0
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.1.24
fonttools==4.55.6
frozenlist==1.5.0
fsspec==2024.12.0
google-api-core==2.24.0
google-auth==2.38.0
google-cloud-vision==3.9.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
html2text==2024.2.26
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.15
langchain-core==0.3.31
langchain-text-splitters==0.3.5
langdetect==1.0.9
langsmith==0.3.1
llama-cloud==0.1.10
llama-index==0.12.14
llama-index-agent-openai==0.4.2
llama-index-cli==0.4.0
llama-index-core==0.12.14
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.4
llama-index-llms-openai==0.3.14
llama-index-multi-modal-llms-openai==0.4.2
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.4
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.20
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
multipart==0.2.4
mypy-extensions==1.0.0
ndjson==0.3.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
openai==0.28.0
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20240706
pdfservices-sdk==2.3.0
pep517==0.13.0
pi_heif==0.21.0
pikepdf==9.5.1
pillow==11.1.0
pip==21.2.3
polling==0.3.2
polling2==0.5.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.4.0
pycocotools==2.0.8
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.14.0
PyJWT==2.4.0
PyMuPDF==1.25.2
pyparsing==3.0.9
pypdf==5.2.0
pypdfium2==4.30.1
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.0
setuptools==57.4.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
sqlparse==0.5.3
striprtf==0.0.26
sympy==1.13.1
tenacity==9.0.0
theflow==0.8.6
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.14
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.1
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2025.1
unstructured==0.16.15
unstructured-client==0.29.0
unstructured-inference==0.8.6
unstructured.pytesseract==0.3.13
urllib3==1.26.13
uuid==1.30
webencodings==0.5.1
wrapt==1.17.2
xlrd==2.0.1
yarl==1.18.3
zstandard==0.23.0
